Factor out reused serverless i18n logic into hook
